Delete BIFM Project Executive Summary-2.docx
EXECUTIVE SUMMARY
We at the Big Store Company work continuously on ways to improve our service to customers. We aim to expand our sales and services, and we would like to grow specifically in the video and audio equipment in the electronics category. We aim to provide good quality products and an excellent retail experience for our audience, for which we aim to focus more on a standardized training approach. It has been seen that a more standardized way of training for an organization's different locations and teams has been very effective.
To account for the basic growth models' behavior, a variety of rationalizations can be applied. The fact that the economy is performing well benefits our ability to accomplish our goals. Unless one or both factors are met, we should be prepared to pursue alternative strategies such as narrowing our focus on a certain category or expanding our online presence. Successfully navigating a competitive marketplace requires consideration and manufacturing audio and video equipment across a diverse range of categories, but our current strategies have been demonstrated to be effective in our position. For instance, Costco is the largest provider of electronics, with a low-price strategy that enables it to obtain high-quality items in bulk through its warehouses without sacrificing quality. This results in a loyal customer base, and the number of customers has increased significantly because of this strategy, as evidenced by the SWOT Analysis.
From our analysis, as shown in the Revenue by Call Center and Category report, it is evident that call centers in New York have very high revenue, with Electronics being the highest contributor, followed by Milwaukee call centers and web. If we take a closer look at the New York call center, out of the total revenue of $7M, nearly $5M of revenue comes from Electronics. Therefore, we aim to increase our sales in this category since our customers already have a good impression of our electronics products. To achieve the 5% growth hike, one of the significant steps can be to start focusing more on products other than video equipment and TVs since they are already performing well on sales. Even if the profit margin is not a lot, concentrating on other products will help us build our market for different products. Electronic products are one of the top sales in the eCommerce platform, and we aim to leverage this in not only New York but also in other locations.  
For our 10% growth plan, one of the key focuses would be improving our online retail service. From the analysis that only 11.4% of our revenue comes from eCommerce. eCommerce provides convenient ways to shop with the comfort of shopping from phones or tablets. One of the most important steps to improve this is to improve our existing user interface with a more customer friendly UX/UI, which lets them compare prices or browse through a variety of different offers very quickly. In addition, targeted sales also help improve sales a lot. We can build and maintain email lists to focus on targeted marketing, meet customers on live sessions on social media, and develop buyer personas to make email marketing more focused.
To meet the 15% goal, we may focus primarily on Electronics as this is where our growth is highly likely to increase. Video equipment and cameras are one of the top subcategories in electronics. More and more video content has been created and watched in recent years. In one of the studies posted by Oberlo, it is stated that 54% prefer video content over any other form of content from a business they like or support. Video marketing strategies have been seen to be highly profitable for a brand. We can use this as our strength in both marketing and increasing video equipment sales.  
One of the few tactics that can be used to overcome potential impediments to success is to expand the distribution channels. We can focus our efforts on product marketing and targeted marketing. This will enable us to build brand awareness for our products and help to enhance the usability of our products.
Every year, income is predicted to grow by a certain percentage. Formulate a strategy for achieving this increase during the next five years. If the baseline growth model, competitive marketplace, products, and strategy are all qualitatively justified in this plan, the plan should be implemented. To keep your business expanding and prospering, hiring an expert to develop your budget plan is an investment and a cost that you must bear as a Big Store Company.
In conclusion, to meet the projected growth targets models we need to follow the following three steps. First step to meet the 5% growth target is the implementation of best practices and standardization throughout the company. This provides consistency and removes waste throughout the organization and strong brand awareness. Second, expand our market share within the northeast region and the web. With a larger market share we can have more leverage and buying power, which allows Big Store Company the ability to provide competitive pricing on our products which is favorable to our customers. The final step is really driving our camera and audio/video equipment and targeting specific customers by developing buyer personas making our email marketing more powerful. 

Reference:
	•	25th, A. G. on O., & Philip Verghese Ariel on October 24th. (2022, February 4). How to increase ecommerce sales: 10 tactics from 53 experts (2018). The BigCommerce Blog. Retrieved March 7, 2022, from https://www.bigcommerce.com/blog/increase-ecommerce-sales/#tactics-to-improve-ecommerce-sales
	•	Doofinder, S. S. E. C. at, Smith, S., & Doofinder, E. C. at. (2021, December 3). Marketing strategies for electronics shops [> 2022 <]. Doofinder. Retrieved March 7, 2022, from https://www.doofinder.com/en/blog/marketing-for-consumer-electronics-shops
	•	Ezeanu, E. (2013, May 7). 5 ways to overcome success barriers. Lifehack. Retrieved March 7, 2022, from https://www.lifehack.org/articles/productivity/5-ways-to-overcome-success-barriers.html
	•	Mohsin, M. (2022, January 16). 10 video marketing statistics you should know for 2022 [infographic]. Oberlo. Retrieved March 7, 2022, from https://www.oberlo.com/blog/video-marketing-statistics